Table Messages renommée Envelop, champ Message renommé Word

+++ backend-spring/src/main/java/com/diefthyntis/chatop/diefthyntis/utils/NumberUtils.java
@@ -8,6 +8,6 @@ public class NumberUtils {
 	    }
 	
 	public static Integer convertToInteger(String data){
-	      return Optional.ofNullable(data).map(Integer::parseInt).orElse(0f);
+	      return Optional.ofNullable(data).map(Integer::parseInt).orElse((int) 0f);
 	    }
 }
